requirements.txt
supabase
openai
python-dotenv
tqdm
tiktoken